@@ -11,7 +11,7 @@
     }
   ],
   "require": {
-    "php": ">=7.1",
+    "php": ">=8.1",
     "ext-calendar": "*",
     "ext-sockets": "*"
   },
@@ -25,6 +25,6 @@
     "psr-4": { "District5Tests\\DateTests\\": "tests/DateTests/" }
   },
   "require-dev": {
-    "phpunit/phpunit": "^8"
+    "phpunit/phpunit": "9.*"
   }
 }
